requirements.txt
Django==3.0.2
psycopg2-binary==2.8.4
python-telegram-bot==12.3.0
XlsxWriter==1.2.7
